move legendre sanity check to iterators module
