+++ src/legendre/mod.rs
@@ -292,26 +292,6 @@ mod tests {
     use approx::assert_abs_diff_eq;
 
     #[test]
-    fn iterator_sanity_check() {
-        for deg in (10..=100).step_by(10) {
-            let rule = GaussLegendre::new(deg);
-            assert_eq!(rule.degree(), deg);
-            for ((ni, wi), (nn, ww)) in rule.iter().zip(rule.nodes().zip(rule.weights())) {
-                assert_abs_diff_eq!(ni, nn);
-                assert_eq!(wi, ww);
-            }
-            for ((ni, wi), (nn, ww)) in rule
-                .as_node_weight_pairs()
-                .iter()
-                .zip(rule.nodes().zip(rule.weights()))
-            {
-                assert_abs_diff_eq!(ni, nn);
-                assert_eq!(wi, ww);
-            }
-        }
-    }
-
-    #[test]
     fn check_degree_3() {
         let rule = GaussLegendre::new(3);
